services/sam/pyproject.toml
[project]
name = "sam-service"
version = "0.1.0"
description = "HTTP service wrapping SAM 3 for the facade-grammar pipeline."
requires-python = ">=3.14"
dependencies = [
    "torch>=2.11,<3",
    "torchvision>=0.22,<1",
    "transformers>=5.5,<6",
    "pillow>=11.0,<12",
    "litserve>=0.2.17,<0.3",
    "pydantic>=2.13,<3",
    "pydantic-settings>=2.13,<3",
    "numpy>=2.2,<3",
]

[project.scripts]
sam-service = "sam_service.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sam_service"]
